App2
The second class of the project. In this program a while loop is used to simulate the number of times a weight is lifted, with each weight reptition being a value of 1. A complete workout is 4 sets of 12, which is 48 repetitions total. Normally while weight lifting after you have completed 12 repetitions you rest for a short period, which is reflected in the second if statement, in that every 12 repitions a rest message will appear. As soon as 48 repetitions in the loop are completed the first if statement outputs the number of repetitions and sets that has been performed and the loop terminates.