App3
In this program a couple items are added to an imaginary weight scale at the same time. The maximum weight the scale can hold is a value of 200 (lbs). After each items weight is added together, the if statement checks to see if the total has reached or exceeded the maximum weight limit represented by the variable maximumWeightLimit, which has a data value of 200. Since in this case it does not reach or exceed the maximum weight, the else statement is used since the if condition has not been met, and a message is outputted to the effect that the weight scale max hasn't been reached.