App4
This program uses the Math class to generate a random number from 0-100, simulating test scores. Then a series of if & else if statements check what the score is, and based upon it's conditions assigns a grade, and outputs a message to that effect. If the score is below 60, the else statement is used for grade of F and outputs a message indicating the student has failed.